requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.13.1
plotly==5.24.1
seaborn==0.13.2
matplotlib==3.9.2
openpyxl==3.1.5
pyarrow==18.1.0
joblib==1.4.2
xgboost==2.1.3
lightgbm==4.5.0
catboost==1.2.7
ydata-profiling==4.12.0
shap==0.46.0
imbalanced-learn==0.12.4
pytest==8.3.4
black==24.10.0
ruff==0.8.4
mypy==1.13.0
pyyaml==6.0.2
